Updated to use updated store functionality.
When the watch launches due to LAUNCH_WAKEUP_PERIOD, we check if
a message was used. If so, we send in the /launch event the new
weight and its corresponding msg_id.

BUG: if the watch has an abnormal launch sequence, it can cause
the launch event to be sent before the weight is updated, causing
the event to not contain the new weight information thus losing the
update.

BUG: if the watch could not send the launch message successfully,
the watch does not save the new weight information and thus is not
replayed during reconnection.